feat(sidebar): add new project hooks for CMS, Companies, CRM, Docs Library, Ecommerce, Human Resources, Program Management, Powers, and Settings
- Implemented `getCmsProject` for CMS sidebar integration.
- Created `getCompaniesProject` for managing company-related sidebar items.
- Added `getCrmProject` to handle CRM functionalities in the sidebar.
- Developed `getDocsLibraryProject` for document library navigation.
- Introduced `getEcommerceProject` for ecommerce-related sidebar links.
- Established `getHumanResourcesProject` for HR management in the sidebar.
- Created `getProgramManagementProject` for program settings access.
- Added `getPowersProject` for managing access to company programs.
- Implemented `getSettingsProject` for user and system settings navigation.
- Updated `useSidebarProjects` to include new project hooks.
- Defined types for sidebar project properties in `types.ts`.
