+++ components/shared/layout/sidebar-content.tsx
@@ -1,21 +1,11 @@
-"use client";
+﻿"use client";
 
 import * as React from "react";
-import {
-  LayoutDashboardIcon,
-  RollerCoasterIcon,
-  UserIcon,
-  Settings,
-  FolderClosed,
-  Inbox,
-  LibraryBig,
-  FileText,
-} from "lucide-react";
+import { Settings, FileText } from "lucide-react";
 import { useTranslations } from "next-intl";
 import { usePathname } from "@i18n/navigation";
 import { ROUTER } from "@/router";
-import SettingsIcon from "@/public/icons/settings";
-import InboxIcon from "@/public/icons/inbox-icon";
+
 import { useSidebarMenu } from "@/hooks/useSidebarMenu";
 import { SUPER_ENTITY_SLUG } from "@/constants/super-entity-slug";
 import { Menu, Project } from "@/types/sidebar-menu";
@@ -23,18 +13,17 @@ import type { Entity } from "@/types/sidebar-menu";
 import { usePermissions } from "@/lib/permissions/client/permissions-provider";
 import { PERMISSIONS } from "@/lib/permissions/permission-names";
 import { createPermissions } from "@/lib/permissions/permission-names/default-permissions";
-import ClipboardClockIcon from "@/public/icons/clipboard-clock";
 import { SidebarProgramsListV2 } from "./sidebar-programs-v2";
 import { UserRoleType } from "@/app/[locale]/(main)/client-profile/[id]/types";
-import { UsersRole } from "@/constants/users-role.enum";
+
 import SidebarFooterContent from "./sidebar-footer-content";
 import { AvatarGroup } from "../avatar-group";
 import { useAuthStore } from "@/modules/auth/store/use-auth";
 import Image from "next/image";
 import LogoPlaceholder from "@/public/images/logo-placeholder-image.png";
 import useCurrentAuthCompany from "@/hooks/use-auth-company";
-import InboxIconWithCount from "@/components/icons/inbox";
 import CrmInboxIconWithCount from "@/components/icons/crm-inbox";
+import { useSidebarProjects } from "@/hooks/use-sidebar-projects";
 
 interface SidebarContentWrapperProps {
   name?: string;
@@ -76,26 +65,6 @@ export function SidebarContentWrapper({
   }, []);
 
   const fullPath = getPathnameWithoutLocale(path);
-
-  const userProfileUrl = React.useMemo(() => {
-    const isEmployee = userTypes.some(
-      (userType) => userType.role == UsersRole.Employee,
-    );
-    const isBroker = userTypes.some(
-      (userType) => userType.role == UsersRole.Broker,
-    );
-    const isClient = userTypes.some(
-      (userType) => userType.role == UsersRole.Client,
-    );
-    if (userTypes.length > 0) {
-      if (isEmployee) {
-        return ROUTER.USER_PROFILE;
-      } else if (isBroker || isClient) {
-        return ROUTER.CLIENT_PROFILE;
-      }
-    }
-    return `${ROUTER.CLIENT_PROFILE}?readonly=true`;
-  }, [userTypes]);
 
   const mergeProjectsAndMenu = React.useCallback(
     function (
@@ -169,8 +138,7 @@ export function SidebarContentWrapper({
             isActive:
               fullPath === ROUTER.CRM.inbox ||
               fullPath.startsWith(`${ROUTER.CRM.inbox}/`),
-            show:
-              !isCentralCompany && can([PERMISSIONS.clientRequest.list]),
+            show: !isCentralCompany && can([PERMISSIONS.clientRequest.list]),
           };
 
           return {
@@ -197,618 +165,10 @@ export function SidebarContentWrapper({
       });
       return formatted;
     },
-    [can, t, pageName, isCentralCompany],
+    [can, t, pageName, fullPath, isCentralCompany],
   );
 
-  const SidebarProjects: Project[] = React.useMemo(() => {
-    const settingsRoutesNames = [
-      ROUTER.SETTINGS,
-      ROUTER.DASHBOARD,
-      ROUTER.CLIENT_PROFILE,
-      ROUTER.USER_PROFILE,
-      ROUTER.COMPANY_PROFILE,
-    ];
-    const rolesObj = {
-      name: t("Sidebar.Roles"),
-      url: ROUTER.ROLES,
-      icon: RollerCoasterIcon,
-      isActive: pageName === ROUTER.ROLES,
-      show: can(Object.values(PERMISSIONS.role)),
-    };
-
-    const permissionsObj = {
-      name: t("Sidebar.Powers"),
-      url: ROUTER.PERMISSIONS,
-      icon: LayoutDashboardIcon,
-      isActive: pageName === ROUTER.PERMISSIONS,
-      show: can(Object.values(PERMISSIONS.permission)),
-    };
-
-    const data: Project[] = [
-      {
-        name: t("Sidebar.WorkPanel"),
-        urls: [
-          ROUTER.WORK_PANEL_SETTINGS,
-          ROUTER.PROJECTS_SETTINGS,
-          ROUTER.ALL_PROJECTS,
-          ROUTER.PROJECTS_INBOX,
-        ],
-        icon: LayoutDashboardIcon,
-        isActive:
-          fullPath.startsWith(ROUTER.WORK_PANEL_SETTINGS) ||
-          fullPath.startsWith(ROUTER.ALL_PROJECTS),
-        slug: "work-panel",
-        show: !isCentralCompany,
-        sub_entities: [
-          {
-            name: t("Sidebar.WorkPanelSettings"),
-            url: ROUTER.WORK_PANEL_SETTINGS,
-            icon: SettingsIcon,
-            isActive:
-              fullPath.startsWith(ROUTER.WORK_PANEL_SETTINGS) ||
-              fullPath.startsWith(ROUTER.PROJECTS_SETTINGS),
-            show: !isCentralCompany,
-          },
-          {
-            name: t("Sidebar.Projects"),
-            url: ROUTER.ALL_PROJECTS,
-            icon: FolderClosed, // Using FolderClosed as it's already imported
-            isActive:
-              fullPath.startsWith("/projects") &&
-              !fullPath.startsWith(ROUTER.PROJECTS_SETTINGS) &&
-              !fullPath.startsWith(ROUTER.PROJECTS_INBOX),
-            show: !isCentralCompany,
-          },
-          {
-            name: t("Sidebar.Inbox"),
-            url: ROUTER.PROJECTS_INBOX,
-            icon: InboxIconWithCount,
-            isActive: fullPath.startsWith(ROUTER.PROJECTS_INBOX),
-            show: !isCentralCompany,
-          },
-        ],
-      },
-      {
-        name: t("Sidebar.Companies"),
-        urls: [ROUTER.COMPANIES],
-        icon: LayoutDashboardIcon,
-        isActive: pageName === ROUTER.COMPANIES,
-        slug: SUPER_ENTITY_SLUG.COMPANY,
-        show: isCentralCompany,
-        sub_entities: [
-          {
-            name: t("Sidebar.CompaniesList"),
-            url: ROUTER.COMPANIES,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.COMPANIES,
-            show:
-              isCentralCompany && can(Object.values(PERMISSIONS.company.list)),
-          },
-        ],
-      },
-      {
-        name: t("Sidebar.Users"),
-        icon: UserIcon,
-        urls: [ROUTER.USERS],
-        isActive: pageName === ROUTER.USERS,
-        slug: SUPER_ENTITY_SLUG.USERS,
-        show: isCentralCompany,
-        sub_entities: [
-          {
-            name: t("Sidebar.UsersList"),
-            url: ROUTER.USERS,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.USERS,
-            show: isCentralCompany && can(Object.values(PERMISSIONS.user.list)),
-          },
-        ],
-      },
-      {
-        name: t("Sidebar.HumanResources"),
-        icon: LayoutDashboardIcon,
-        urls: [ROUTER.Organizational_Structure, ROUTER.WORK_PANEL],
-        isActive: pageName === ROUTER.Organizational_Structure,
-        slug: SUPER_ENTITY_SLUG.HRM,
-        show: !isCentralCompany,
-        sub_entities: [
-          {
-            name: t("WorkPanel.title"),
-            url: ROUTER.WORK_PANEL,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.WORK_PANEL,
-            show: !isCentralCompany,
-          },
-          {
-            name: t("Sidebar.OrganizationalStructure"),
-            url: ROUTER.Organizational_Structure,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.Organizational_Structure,
-            show:
-              !isCentralCompany &&
-              can([
-                PERMISSIONS.organization.branch.view,
-                PERMISSIONS.organization.department.view,
-                PERMISSIONS.organization.jobTitle.list,
-                PERMISSIONS.organization.jobType.list,
-                PERMISSIONS.organization.management.view,
-                PERMISSIONS.organization.users.view,
-              ]),
-          },
-          {
-            name: t("Sidebar.AttendanceDeparture"),
-            url: ROUTER.AttendanceDeparture,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.AttendanceDeparture,
-            show:
-              !isCentralCompany &&
-              can([PERMISSIONS.attendance.attendance_departure.view]),
-          },
-          {
-            name: t("Sidebar.HRSettings"),
-            url: ROUTER.HR_SETTINGS,
-            icon: SettingsIcon,
-            isActive: pageName === ROUTER.HR_SETTINGS,
-            show:
-              !isCentralCompany && can([PERMISSIONS.attendance.settings.view]),
-          },
-        ],
-      },
-      {
-        name: t("Sidebar.ProgramManagement"),
-        slug: SUPER_ENTITY_SLUG.PM,
-        icon: LayoutDashboardIcon,
-        urls: [ROUTER.PROGRAM_SETTINGS.USERS],
-        isActive: pageName === ROUTER.PROGRAM_SETTINGS.USERS,
-        show: isCentralCompany,
-        sub_entities: [
-          {
-            name: t("Sidebar.Users"),
-            url: ROUTER.PROGRAM_SETTINGS.USERS,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.PROGRAM_SETTINGS.USERS,
-            show: isCentralCompany && can(Object.values(PERMISSIONS.subEntity)),
-          },
-        ],
-      },
-      {
-        name: t("Sidebar.CRM"),
-        slug: SUPER_ENTITY_SLUG.CRM,
-        icon: LayoutDashboardIcon,
-        urls: [
-          ROUTER.CRM.clients,
-          ROUTER.CRM.brokers,
-          ROUTER.CRM.settings,
-          ROUTER.CRM.pricesOffers,
-          ROUTER.CRM.clientRequests,
-          ROUTER.CRM.inbox,
-        ],
-        isActive: [
-          ROUTER.CRM.clients,
-          ROUTER.CRM.brokers,
-          ROUTER.CRM.settings,
-          ROUTER.CRM.pricesOffers,
-          ROUTER.CRM.clientRequests,
-          ROUTER.CRM.inbox,
-        ].some((route) => path === route || path.endsWith(route)),
-        show: !isCentralCompany,
-        sub_entities: [
-          {
-            name: t("Sidebar.PricesOffers"),
-            url: ROUTER.CRM.pricesOffers,
-            icon: FileText,
-            isActive: fullPath === ROUTER.CRM.pricesOffers,
-            show: can([PERMISSIONS.crm.pricesOffers.list]),
-          },
-        ],
-      },
-      {
-        name: t("Sidebar.docs-library"),
-        slug: SUPER_ENTITY_SLUG.DOCS_LIBRARY,
-        icon: LibraryBig,
-        urls: [ROUTER.DOCS_LIBRARY],
-        isActive: pageName === ROUTER.DOCS_LIBRARY,
-        show: !isCentralCompany,
-        sub_entities: [
-          {
-            name: t("Sidebar.docs-library-docs"),
-            url: ROUTER.DOCS_LIBRARY,
-            icon: FolderClosed,
-            isActive: pageName === ROUTER.DOCS_LIBRARY,
-            show: !isCentralCompany && can([PERMISSIONS.library.folder.list]),
-          },
-        ],
-      },
-      {
-        name: t("Sidebar.Settings"),
-        icon: SettingsIcon,
-        isActive: settingsRoutesNames.indexOf(pageName) !== -1,
-        slug: SUPER_ENTITY_SLUG.SETTINGS,
-        urls: [userProfileUrl, ROUTER.COMPANY_PROFILE, ROUTER.SETTINGS],
-        show: true,
-        sub_entities: [
-          {
-            name: t("Sidebar.UserProfileSettings"),
-            url: userProfileUrl,
-            icon: UserIcon,
-            isActive: pageName === userProfileUrl,
-            show: can([
-              ...Object.values(PERMISSIONS.userProfile).flatMap((p) =>
-                Object.values(p),
-              ),
-              ...Object.values(PERMISSIONS.profile).flatMap((p) =>
-                Object.values(p),
-              ),
-            ]),
-          },
-          {
-            name: t("Sidebar.DashboardSettings"),
-            url: ROUTER.COMPANY_PROFILE,
-            icon: InboxIcon,
-            isActive: pageName === ROUTER.COMPANY_PROFILE,
-            show: can(
-              Object.values(PERMISSIONS.companyProfile).flatMap((p) =>
-                Object.values(p),
-              ),
-            ),
-          },
-          {
-            name: t("Sidebar.SystemSettings"),
-            url: ROUTER.SETTINGS,
-            icon: InboxIcon,
-            isActive: pageName === ROUTER.SETTINGS,
-            show: can([
-              PERMISSIONS.identifier.list,
-              PERMISSIONS.loginWay.list,
-              PERMISSIONS.driver.view,
-            ]),
-          },
-          {
-            name: t("Sidebar.ActivitiesLogs"),
-            url: ROUTER.ACTIVITIES_LOGS,
-            icon: ClipboardClockIcon,
-            isActive: pageName === ROUTER.ACTIVITIES_LOGS,
-            show: can([PERMISSIONS.activityLogs.list]),
-          },
-          rolesObj,
-          permissionsObj,
-        ],
-      },
-      {
-        name: t("Sidebar.Powers"),
-        icon: SettingsIcon,
-        isActive: pageName === ROUTER.Powers,
-        slug: SUPER_ENTITY_SLUG.POWERS,
-        urls: [ROUTER.Programs],
-        show: isCentralCompany,
-        sub_entities: [
-          {
-            name: t("Sidebar.PackagesAndPrograms"),
-            url: ROUTER.Programs,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.Programs,
-            show:
-              isCentralCompany &&
-              can(Object.values(PERMISSIONS.companyAccessProgram)),
-          },
-        ],
-      },
-      {
-        name: t("Sidebar.ecommerce"),
-        icon: SettingsIcon,
-        isActive: [
-          ROUTER.HomeStore,
-          ROUTER.Products,
-          ROUTER.Brands,
-          ROUTER.Categories,
-          ROUTER.Terms,
-          ROUTER.warehouse,
-          ROUTER.Discounts,
-          ROUTER.Coupons,
-        ].includes(pageName),
-        show: !isCentralCompany,
-        slug: SUPER_ENTITY_SLUG.ECOMMERCE,
-        urls: [
-          ROUTER.HomeStore,
-          ROUTER.Products,
-          ROUTER.Brands,
-          ROUTER.Categories,
-          ROUTER.Terms,
-          ROUTER.warehouse,
-          ROUTER.Discounts,
-          ROUTER.Coupons,
-          ROUTER.PaymentMethods,
-        ],
-        sub_entities: [
-          {
-            name: t("Sidebar.HomeStore"),
-            url: ROUTER.HomeStore,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.HomeStore,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.ecommerce.dashboard)),
-          },
-          {
-            name: t("Sidebar.Products"),
-            url: ROUTER.Products,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.Products,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.ecommerce.product)),
-          },
-          {
-            name: t("Sidebar.Requests"),
-            url: ROUTER.requests,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.requests,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.ecommerce.order)),
-          },
-          {
-            name: t("Sidebar.Categories"),
-            url: ROUTER.Categories,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.Categories,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.ecommerce.category)),
-          },
-          {
-            name: t("Sidebar.Brands"),
-            url: ROUTER.Brands,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.Brands,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.ecommerce.brand)),
-          },
-          {
-            name: t("Sidebar.Coupons"),
-            url: ROUTER.Coupons,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.Coupons,
-            show:
-              !isCentralCompany &&
-              can([
-                PERMISSIONS.ecommerce.coupon.list,
-                PERMISSIONS.ecommerce.featureDeal.list,
-                PERMISSIONS.ecommerce.flashDeal.list,
-                PERMISSIONS.ecommerce.dealDay.list,
-              ]),
-          },
-          {
-            name: t("Sidebar.Warehouse"),
-            url: ROUTER.warehouse,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.warehouse,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.ecommerce.warehouse)),
-          },
-          {
-            name: t("Sidebar.PagesSettings"),
-            url: ROUTER.pagesSettings,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.pagesSettings,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.ecommerce.banner)),
-          },
-          {
-            name: t("Sidebar.Terms"),
-            url: ROUTER.Terms,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.Terms,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.ecommerce.page)),
-          },
-          {
-            name: t("Sidebar.SocialMedia"),
-            url: ROUTER.SocialMedia,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.SocialMedia,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.ecommerce.socialMedia)),
-          },
-          {
-            name: t("Sidebar.PaymentMethods"),
-            url: ROUTER.PaymentMethods,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.PaymentMethods,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.ecommerce.paymentMethod)),
-          },
-        ],
-      },
-      {
-        name: t("Sidebar.CMS.title"),
-        icon: SettingsIcon,
-        isActive: [
-          ROUTER.CMS.CATEGORIES,
-          ROUTER.CMS.ICONS,
-          ROUTER.CMS.MAIN_SETTINGS,
-          ROUTER.CMS.ABOUT_SETTINGS,
-          ROUTER.CMS.TERMS_CONDITIONS,
-          ROUTER.CMS.MAIN_DATA,
-          ROUTER.CMS.FOUNDER,
-          ROUTER.CMS.SERVICES,
-          ROUTER.CMS.OUR_SERVICES,
-          ROUTER.CMS.NEWS,
-          ROUTER.CMS.THEMES,
-          ROUTER.CMS.COMMUNICATION_MESSAGES,
-          ROUTER.CMS.PROJECTS,
-          ROUTER.CMS.THEME_SETTING,
-        ].includes(pageName),
-        show: !isCentralCompany,
-        slug: SUPER_ENTITY_SLUG.CMS,
-        urls: [
-          ROUTER.CMS.CATEGORIES,
-          ROUTER.CMS.ICONS,
-          ROUTER.CMS.MAIN_SETTINGS,
-          ROUTER.CMS.ABOUT_SETTINGS,
-          ROUTER.CMS.TERMS_CONDITIONS,
-          ROUTER.CMS.MAIN_DATA,
-          ROUTER.CMS.FOUNDER,
-          ROUTER.CMS.SERVICES,
-          ROUTER.CMS.OUR_SERVICES,
-          ROUTER.CMS.NEWS,
-          ROUTER.CMS.PROJECTS,
-          ROUTER.CMS.THEME_SETTING,
-          ROUTER.CMS.THEMES,
-          ROUTER.CMS.COMMUNICATION_MESSAGES,
-        ],
-        sub_entities: [
-          {
-            name: t("Sidebar.CMS.Themes"),
-            url: ROUTER.CMS.THEMES,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.CMS.THEMES,
-            show:
-              !isCentralCompany && can(Object.values(PERMISSIONS.CMS.themes)),
-          },
-          {
-            name: t("Sidebar.CMS.ThemeSetting"),
-            url: ROUTER.CMS.THEME_SETTING,
-            icon: SettingsIcon,
-            isActive: pageName === ROUTER.CMS.THEME_SETTING,
-            show:
-              !isCentralCompany &&
-              can([
-                PERMISSIONS.CMS.themeSetting.view,
-                PERMISSIONS.CMS.themeSetting.update,
-              ]),
-          },
-          {
-            name: t("Sidebar.CMS.Categories"),
-            url: ROUTER.CMS.CATEGORIES,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.CMS.CATEGORIES,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.CMS.categories)),
-          },
-          {
-            name: t("Sidebar.CMS.Icons"),
-            url: ROUTER.CMS.ICONS,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.CMS.ICONS,
-            show:
-              !isCentralCompany && can(Object.values(PERMISSIONS.CMS.icons)),
-          },
-          {
-            name: t("Sidebar.CMS.Services"),
-            url: ROUTER.CMS.SERVICES,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.CMS.SERVICES,
-            show:
-              !isCentralCompany &&
-              can([
-                PERMISSIONS.CMS.services.list,
-                PERMISSIONS.CMS.services.update,
-              ]),
-          },
-          {
-            name: t("Sidebar.CMS.Projects"),
-            url: ROUTER.CMS.PROJECTS,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.CMS.PROJECTS,
-            show:
-              !isCentralCompany && can(Object.values(PERMISSIONS.CMS.projects)),
-          },
-          {
-            name: t("Sidebar.CMS.News"),
-            url: ROUTER.CMS.NEWS,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.CMS.NEWS,
-            show: !isCentralCompany && can(Object.values(PERMISSIONS.CMS.news)),
-          },
-          {
-            name: t("Sidebar.CMS.Founder"),
-            url: ROUTER.CMS.FOUNDER,
-            icon: UserIcon,
-            isActive: pageName === ROUTER.CMS.FOUNDER,
-            show:
-              !isCentralCompany && can(Object.values(PERMISSIONS.CMS.founder)),
-          },
-          {
-            name: t("Sidebar.CMS.MainSettings"),
-            url: ROUTER.CMS.MAIN_SETTINGS,
-            icon: SettingsIcon,
-            isActive: pageName === ROUTER.CMS.MAIN_SETTINGS,
-            show:
-              !isCentralCompany &&
-              can([
-                PERMISSIONS.CMS.mainSettings.view,
-                PERMISSIONS.CMS.mainSettings.update,
-              ]),
-          },
-          {
-            name: t("Sidebar.CMS.CommunicationMessages"),
-            url: ROUTER.CMS.COMMUNICATION_MESSAGES,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.CMS.COMMUNICATION_MESSAGES,
-            show:
-              !isCentralCompany &&
-              can(Object.values(PERMISSIONS.CMS.communicationContactMessages)),
-          },
-          {
-            name: t("Sidebar.CMS.CommunicationSettings"),
-            url: ROUTER.CMS.COMMUNICATION_SETTINGS,
-            icon: SettingsIcon,
-            isActive: pageName === ROUTER.CMS.COMMUNICATION_SETTINGS,
-            show:
-              !isCentralCompany &&
-              can([
-                PERMISSIONS.CMS.aboutSetting.update,
-                PERMISSIONS.CMS.aboutSetting.view,
-              ]),
-          },
-          {
-            name: t("Sidebar.CMS.AboutUsPage"),
-            url: ROUTER.CMS.ABOUT_SETTINGS,
-            icon: SettingsIcon,
-            isActive: pageName === ROUTER.CMS.ABOUT_SETTINGS,
-            show:
-              !isCentralCompany &&
-              can([
-                PERMISSIONS.CMS.aboutSetting.update,
-                PERMISSIONS.CMS.aboutSetting.view,
-              ]),
-          },
-          {
-            name: t("Sidebar.CMS.OurServices"),
-            url: ROUTER.CMS.OUR_SERVICES,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.CMS.OUR_SERVICES,
-            show:
-              !isCentralCompany &&
-              can([
-                PERMISSIONS.CMS.services.list,
-                PERMISSIONS.CMS.services.update,
-              ]),
-          },
-          {
-            name: t("Sidebar.CMS.TermsConditions"),
-            url: ROUTER.CMS.TERMS_CONDITIONS,
-            icon: LayoutDashboardIcon,
-            isActive: pageName === ROUTER.CMS.TERMS_CONDITIONS,
-            show:
-              !isCentralCompany &&
-              can([
-                PERMISSIONS.CMS.termsConditions.view,
-                PERMISSIONS.CMS.termsConditions.update,
-              ]),
-          },
-        ],
-      },
-    ];
-    return data;
-  }, [pageName, path, isCentralCompany, can, t, userProfileUrl]);
+  const SidebarProjects = useSidebarProjects(userTypes);
 
   const all = React.useMemo(() => {
     if (isLoading) return [];
